Update README.md
This repository contains a machine learning project focused on email spam detection. The goal is to build a model capable of classifying emails as either spam or not spam (ham). The project showcases the application of machine learning techniques in filtering unwanted or malicious emails.

Dataset
The dataset used in this project consists of a collection of labeled emails, with each email marked as spam or ham. The features include the email content, sender information, and other relevant attributes.